Rearrange logs, move items to lib, simulate sync
